refactor(pricing-service): replace old getFiatValue with BN-based implementation
